+++ elasticsearch-model/examples/activerecord_associations.rb
@@ -80,7 +80,7 @@ end
 
 # ----- Elasticsearch client setup ----------------------------------------------------------------
 
-Elasticsearch::Model.client = Elasticsearch::Client.new log: true
+Elasticsearch::Model.client = OpenSearch::Client.new log: true
 Elasticsearch::Model.client.transport.transport.logger.formatter = proc { |s, d, p, m| "\e[2m#{m}\n\e[0m" }
 
 # ----- Search integration ------------------------------------------------------------------------