+++ elasticsearch-model/examples/activerecord_mapping_edge_ngram.rb
@@ -68,7 +68,7 @@ class Article < ActiveRecord::Base
   end
 end
 
-Article.__elasticsearch__.client = Elasticsearch::Client.new log: true
+Article.__elasticsearch__.client = OpenSearch::Client.new log: true
 
 # Create index