+++ elasticsearch-model/examples/couchbase_article.rb
@@ -66,7 +66,7 @@ Article.create id: '3', title: 'Foo Foo'  rescue nil
 
 # Index data into Elasticsearch
 #
-client = Elasticsearch::Client.new log:true
+client = OpenSearch::Client.new log:true
 
 client.indices.delete index: 'articles' rescue nil
 client.bulk index: 'articles',