+++ elasticsearch-model/examples/mongoid_article.rb
@@ -38,7 +38,7 @@ Moped.logger.level = Logger::DEBUG
 
 Mongoid.connect_to 'articles'
 
-Elasticsearch::Model.client = Elasticsearch::Client.new host: 'localhost:9200', log: true
+Elasticsearch::Model.client = OpenSearch::Client.new host: 'localhost:9200', log: true
 
 class Article
   include Mongoid::Document
@@ -66,7 +66,7 @@ Article.create id: '3', title: 'Foo Foo'
 
 # Index data
 #
-client = Elasticsearch::Client.new host:'localhost:9200', log:true
+client = OpenSearch::Client.new host:'localhost:9200', log:true
 
 client.indices.delete index: 'articles' rescue nil
 client.bulk index: 'articles',