+++ elasticsearch-persistence/spec/repository_spec.rb
@@ -104,7 +104,7 @@ describe Elasticsearch::Persistence::Repository do
       end
 
       it 'sets a default client' do
-        expect(repository.client).to be_a(Elasticsearch::Client)
+        expect(repository.client).to be_a(OpenSearch::Client)
       end
 
 
@@ -120,7 +120,7 @@ describe Elasticsearch::Persistence::Repository do
     context 'when options are provided' do
 
       let(:client) do
-        Elasticsearch::Client.new
+        OpenSearch::Client.new
       end
 
       let(:repository) do
@@ -465,7 +465,7 @@ describe Elasticsearch::Persistence::Repository do
       end
 
       it 'sets a default on the instance' do
-        expect(RepositoryWithoutDSL.new.client).to be_a(Elasticsearch::Client)
+        expect(RepositoryWithoutDSL.new.client).to be_a(OpenSearch::Client)
       end
 
       it 'allows the value to be overridden with options on the instance' do
@@ -562,7 +562,7 @@ describe Elasticsearch::Persistence::Repository do
             it 'raises an error' do
               expect {
                 repository.create_index!
-              }.to raise_exception(Elasticsearch::Transport::Transport::Errors::BadRequest)
+              }.to raise_exception(OpenSearch::Transport::Transport::Errors::BadRequest)
             end
           end
         end