+++ elasticsearch-rails/lib/rails/templates/03-expert.rb
@@ -281,7 +281,7 @@ create_file 'config/initializers/elasticsearch.rb', <<-CODE
 # Connect to specific Elasticsearch cluster
 ELASTICSEARCH_URL = ENV['ELASTICSEARCH_URL'] || 'http://localhost:9200'
 
-Elasticsearch::Model.client = Elasticsearch::Client.new host: ELASTICSEARCH_URL
+Elasticsearch::Model.client = OpenSearch::Client.new host: ELASTICSEARCH_URL
 
 # Print Curl-formatted traces in development into a file
 #