+++ elasticsearch-rails/spec/spec_helper.rb
@@ -35,7 +35,7 @@ RSpec.configure do |config|
     require 'ansi'
     tracer = ::Logger.new(STDERR)
     tracer.formatter = lambda { |s, d, p, m| "#{m.gsub(/^.*$/) { |n| '   ' + n }.ansi(:faint)}\n" }
-    Elasticsearch::Model.client = Elasticsearch::Client.new host: ELASTICSEARCH_URL,
+    Elasticsearch::Model.client = OpenSearch::Client.new host: ELASTICSEARCH_URL,
                                                             tracer: (ENV['QUIET'] ? nil : tracer)
     puts "Elasticsearch Version: #{Elasticsearch::Model.client.info['version']}"